requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
pypdf
python-docx
python-dotenv
pyyaml
streamlit
tenacity
pytest